codepile/enwikibooks/preprocess/requirements.txt
# git+https://github.com/seomoz/simhash-py.git
simhash-py
datasets==2.5.2
emoji==1.7.0
fasttext==0.9.2
fsspec==2022.8.2
gcsfs==2022.8.2
kenlm==0.0.0
numpy==1.21.6
pandas==1.3.5
pypi_kenlm==0.1.20220713
PyYAML==6.0
regex==2022.9.13
sentencepiece==0.1.97
streamlit==1.13.0
tqdm==4.64.1
typer==0.6.1